Config/CLI Maintenance (#342)
- Adds `--nomdns` to disable p2p mdns discovery
- Simplifies flag parsing by utilizing default values
- CLI testing refactored to be more modular
- Makes default config immutable
- Reduce log bloat on startup failure

+++ cmd/netsim/sim.go
@@ -50,7 +50,7 @@ func NewSimulator(num int) (sim *Simulator, err error) {
 			BootstrapNodes: []string{
 				bootnodeAddr,
 			},
-			Port: 5001 + i,
+			Port: 5001 + uint32(i),
 		}
 		sim.Nodes[i] = new(p2p.Service)
 		sim.Nodes[i], err = p2p.NewService(conf, nil)
